mammo-scan-pipeline/requirements.txt
# DICOM Processing Pipeline + PyTorch Requirements
# ===============================================
numpy>=1.21.0
scikit-build
cmake
setuptools>=42
wheel

# DICOM Compression Support
pylibjpeg>=1.4.0,<2.0
pylibjpeg-libjpeg>=1.3.0,<2.0
pylibjpeg-openjpeg>=1.3.0,<2.0
pylibjpeg-rle>=1.3.0

# PyTorch y Deep Learning (NUEVO)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0
pytorch-tabular>=1.0.0       # Para modelos tabulares con PyTorch

# Computer Vision y Medical Imaging (NUEVO)
monai>=1.3.0
albumentations>=1.3.0

# Machine Learning adicionales (NUEVO)
optuna>=3.4.0                # Optimización de hiperparámetros
wandb>=0.15.0                # Tracking de experimentos
tensorboard>=2.14.0          # Visualización de métricas
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0     # Para SMOTE y manejo de clases desbalanceadas
xgboost>=2.0.0               # XGBoost para gradient boosting
networkx>=3.0                # Para análisis de grafos (usado en dicom_image_processing)

# Configuración y utilidades (NUEVO)
pyyaml>=6.0
tqdm>=4.66.0
rich>=13.5.0                 # Para outputs más bonitos
click>=8.1.0                 # Para CLI mejorado

# Jupyter (NUEVO - para notebooks)
jupyter>=1.0.0
ipywidgets>=8.1.0

# Core scientific computing
pandas>=1.3.0
matplotlib>=3.4.0

# Medical image processing
pydicom>=2.3.0
SimpleITK>=2.1.0

# Computer vision and image processing
opencv-python>=4.5.0
scikit-image>=0.18.0

# Radiomics feature extraction
pyradiomics>=3.0.0

# System monitoring and parallel processing
psutil>=5.8.0
joblib>=1.0.0

# Additional utilities
pathlib2>=2.3.6  # For older Python versions
argparse  # Usually built-in, but listed for completeness

# Optional: For Jupyter notebook compatibility
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Development and testing (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
